apps/server/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
asyncpg==0.30.0
pydantic==2.11.3
python-dotenv==1.1.0
